2016.11.13
TpFiles.pm - Changed the FileType() subroutine interface such that only
the first line is scanned and that it will ignore preceding blank lines.

vgmfilerouter-wrapper.sh - To fix the following: everytime the main
script is called, the parameters are appended, such that, it will
continuously add the parameters at the end.